+++ AppLovin/CHANGELOG.md
@@ -1,4 +1,19 @@
 ## Changelog
+* 10.1.2.0
+    * This version of the adapters has been certified with AppLovin 10.1.2 and MoPub SDK 5.16.4.
+
+* 10.1.1.0
+    * This version of the adapters has been certified with AppLovin 10.1.1 and MoPub SDK 5.16.3.
+
+* 10.0.1.0
+    * This version of the adapters has been certified with AppLovin 10.0.1 and MoPub SDK 5.16.0.
+
+* 9.15.3.0
+    * This version of the adapters has been certified with AppLovin 9.15.3 and MoPub SDK 5.16.0.
+
+* 9.15.2.0
+    * This version of the adapters has been certified with AppLovin 9.15.2 and MoPub SDK 5.15.0.
+
 * 9.15.1.0
     * This version of the adapters has been certified with AppLovin 9.15.1 and MoPub SDK 5.15.0.
 
@@ -212,3 +227,5 @@
 	
   * Initial Commit
   	* Adapters moved from [mopub-android-sdk](https://github.com/mopub/mopub-android-sdk) to [mopub-android-mediation](https://github.com/mopub/mopub-android-mediation/)
+      
+